+++ services/web-app/components/markdown/index.js
@@ -1,0 +1,17 @@
+/*
+ * Copyright IBM Corp. 2022, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+export { default as Blockquote } from './blockquote'
+export { default as H1 } from './h1'
+export { default as H2 } from './h2'
+export { default as H3 } from './h3'
+export { default as H4 } from './h4'
+export { default as H5 } from './h5'
+export { default as H6 } from './h6'
+export { default as LI } from './li'
+export { default as OL } from './ol'
+export { default as P } from './p'
+export { default as UL } from './ul'